pyproject.toml
# --- Секция с метаданными для Poetry ---
[tool.poetry]
name = "among-us-bot"
version = "0.1.0"
description = "Play among us with your company offline"
authors = ["n.khabarov <your@email.com>"]
readme = "README.md"


packages = [
    { include = "src" },
    { include = "configs" },
    { include = "ui" }
]

# --- Секция с зависимостями ---
[tool.poetry.dependencies]
python = ">=3.12,<3.15"
pandas = ">=2.3.2,<3.0.0"
numpy = ">=2.3.3,<3.0.0"
openpyxl = ">=3.1.5,<4.0.0"

# --- Эта секция остается без изменений ---
ipykernel = "^6.30.1"
python-telegram-bot = "^22.4"
dotenv = "^0.9.9"
asyncio = "^4.0.0"
aiogram = "^3.23.0"


[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"